Refactor sensor
* Add NOSONAR lines on the go
* Extract DefaultJavaFileScannerContext as top class
