+++ java-checks-testkit/src/main/java/org/sonar/java/checks/verifier/JavaCheckVerifier.java
@@ -35,6 +35,7 @@ import org.fest.assertions.Fail;
 import org.sonar.java.AnalyzerMessage;
 import org.sonar.java.ast.JavaAstScanner;
 import org.sonar.java.ast.visitors.SubscriptionVisitor;
+import org.sonar.java.model.DefaultJavaFileScannerContext;
 import org.sonar.java.model.VisitorsBridge;
 import org.sonar.plugins.java.api.JavaFileScanner;
 import org.sonar.plugins.java.api.JavaFileScannerContext;
@@ -277,7 +278,7 @@ public class JavaCheckVerifier extends SubscriptionVisitor {
   public void scanFile(JavaFileScannerContext context) {
     expected.clear();
     super.scanFile(context);
-    VisitorsBridge.DefaultJavaFileScannerContext djfsc = (VisitorsBridge.DefaultJavaFileScannerContext) context;
+    DefaultJavaFileScannerContext djfsc = (DefaultJavaFileScannerContext) context;
     // leave file.
     checkIssues(djfsc.sourceFile);
     expected.clear();
